+++ extensions/vscode-aml/src/web/extension.ts
@@ -1,8 +1,9 @@
 import vscode, {ExtensionContext, TextEditor} from "vscode";
 import {newAml} from "./new";
 import {convertAmlToDialect, convertJsonToAml, convertSqlToAml} from "./convert";
-import {AmlDocumentSymbolProvider} from "./symbols";
 import {previewAml} from "./preview";
+import {startDiagnostics} from "./diagnostics";
+import {AmlDocumentSymbolProvider} from "./symbols";
 
 export function activate(context: ExtensionContext) {
     context.subscriptions.push(
@@ -11,6 +12,8 @@ export function activate(context: ExtensionContext) {
         vscode.commands.registerTextEditorCommand('aml.fromSQL', (editor: TextEditor) => convertSqlToAml(editor)),
         vscode.commands.registerTextEditorCommand('aml.convert', (editor: TextEditor) => convertAmlToDialect(editor)),
         vscode.commands.registerTextEditorCommand('aml.preview', (editor: TextEditor) => previewAml(editor, context)),
+        startDiagnostics(context),
+        // vscode.languages.registerCompletionItemProvider({language: 'aml'}, new AmlCompletionItemProvider(), ' ', '(', '{', ',', '.'),
         vscode.languages.registerDocumentSymbolProvider({language: 'aml'}, new AmlDocumentSymbolProvider()),
     )
 }